refactor: Rename utils mod to functions
