+++ src/contexts/mod.rs
@@ -1,3 +1,2 @@
-pub mod top_level;
 pub mod git;
 pub mod rbenv;
