+++ src/engine.rs
@@ -1,16 +1,11 @@
-use contexts::top_level::TopLevelContext;
 use parser;
 use eval;
 
-pub struct Engine {
-    top_level_context: TopLevelContext,
-}
+pub struct Engine;
 
 impl Engine {
     pub fn new() -> Engine {
-        Engine {
-            top_level_context: TopLevelContext::new(),
-        }
+        Engine
     }
 
     pub fn run(&self, input: &str) -> String {
